Move platform-specific code for sky/shell into sky/shell/platform
This patch prepares us to add support for Mozart.
