+++ sky/shell/platform/android/AndroidManifest.xml
