+++ sky/shell/platform/android/org/domokit/sky/shell/ResourceCleaner.java
