+++ sky/shell/platform/android/org/domokit/sky/shell/SkyActivity.java
