+++ sky/shell/platform/android/update_service_android.cc
@@ -2,7 +2,7 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-#include "sky/shell/android/update_service_android.h"
+#include "sky/shell/platform/android/update_service_android.h"
 
 #include "base/logging.h"
 #include "base/task_runner_util.h"
