+++ sky/shell/platform/ios/Entitlements.xcent
