+++ sky/shell/platform/ios/Info.plist
