+++ sky/shell/platform/ios/LaunchScreen.storyboardc/Info.plist
