+++ sky/shell/platform/ios/sky_app_delegate.mm
