+++ sky/shell/platform/mac/.gitignore
