+++ sky/shell/platform/mac/Info.plist
