+++ sky/shell/platform/mac/main_mac.mm
@@ -7,8 +7,8 @@
 #include "base/bind.h"
 #include "base/command_line.h"
 #include "base/message_loop/message_loop.h"
-#include "sky/shell/mac/platform_mac.h"
-#include "sky/shell/mac/sky_application.h"
+#include "sky/shell/platform/mac/platform_mac.h"
+#include "sky/shell/platform/mac/sky_application.h"
 #include "sky/shell/switches.h"
 #include "sky/shell/testing/testing.h"
