+++ sky/shell/platform/mac/platform_view_mac.mm
@@ -2,7 +2,7 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-#include "sky/shell/mac/platform_view_mac.h"
+#include "sky/shell/platform/mac/platform_view_mac.h"
 
 namespace sky {
 namespace shell {
