+++ sky/shell/platform/mac/sky_application.mm
@@ -2,7 +2,7 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-#include "sky/shell/mac/sky_application.h"
+#include "sky/shell/platform/mac/sky_application.h"
 
 #include "base/auto_reset.h"
 #include "base/logging.h"
