+++ sky/shell/platform/mac/sky_window.h
